% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMF_maker.R
\name{IMF_maker}
\alias{IMF_maker}
\title{Makes Ensemble Empiric Mode Decomposition  for a given Time Series}
\usage{
IMF_maker(ts, nb_imf)
}
\arguments{
\item{ts}{time series in xts format}

\item{nb_imf}{chosen number of IMFs}
}
\value{
data frame containing the IMF s of a given time series
}
\description{
Makes Ensemble Empiric Mode Decomposition  for a given Time Series
}
\examples{
VVD::IMF_maker(data(a10),6)


}
