% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMF_multi_plot.R
\name{IMF_multi_plot}
\alias{IMF_multi_plot}
\title{Combines all IMF s plots into one graph}
\usage{
IMF_multi_plot(ts)
}
\arguments{
\item{ts}{given time series}
}
\value{
combined plots
}
\description{
function that plots all IMFs into one combined graph
}
\examples{
library(fpp)
data(a10)
VVD::IMF_multi_plot(a10)

}
