% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMF_plot.R
\name{IMF_plot}
\alias{IMF_plot}
\title{Plots a chosen IMF of an EEMD time series}
\usage{
IMF_plot(data, imf)
}
\arguments{
\item{x}{data frame containing IMF s}

\item{i}{the index of the column you want to plot}
}
\value{

}
\description{
Plots a chosen IMF of an EEMD time series
}
\examples{
library(fpp)
data(a10)
#' VVD::IMF_plot(VVD::IMF_maker(a10),VVD::IMF_number(a10),2)

}
