% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMF_SCA.R
\name{IMF_SCA}
\alias{IMF_SCA}
\title{Testing the sationarity of a given time series using EEMD method}
\usage{
IMF_SCA(ts, imf_nb, imf_chosen, beta)
}
\arguments{
\item{ts}{time series data in xts format}

\item{imf_nb}{number  of imfs of EEMD}

\item{imf_chosen}{chosen imf to be tested}

\item{beta}{framing value \link{1-beta , 1+beta}}
}
\value{
a boolean True if test is valid
}
\description{
Testing the sationarity of a given time series using EEMD method
}
\examples{
library(fpp)
data(a10)
VVD::IMF_SCA(a10,6,2,0.3)

}
