% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMF_season.R
\name{IMF_season}
\alias{IMF_season}
\title{Funds the seasonality of the a given time series}
\usage{
IMF_season(ts, pl)
}
\arguments{
\item{ts}{given time series in xts}

\item{plot}{boolean that returns a plot of the seasonality if true is given}
}
\value{
seasonality of the a given time series
}
\description{
Funds the seasonality of the a given time series
}
\examples{
VVD::IMF_season(data(a10),T)

}
