% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMF_trend.R
\name{IMF_trend}
\alias{IMF_trend}
\title{Fonds the trend of a given time series}
\usage{
IMF_trend(ts, pl)
}
\arguments{
\item{ts}{time series xts format}

\item{plot}{boolean if True Plots the trend}
}
\value{
returns an xts variable containing a
}
\description{
Fonds the trend of a given time series
}
\examples{
library(fpp)
data(a10)
VVD::IMF_trend(a10,TRUE)

}
