% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeR.R
\name{modeR}
\alias{modeR}
\title{Mode of of integers vector}
\usage{
modeR(v)
}
\arguments{
\item{v}{vector of integers}
}
\value{
integer with the mode value of the given vector
}
\description{
Mode of of integers vector
}
\examples{
modeR(rpois(n = 50, lambda = 10))
}
