% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMF_plot.R
\name{IMF_plot}
\alias{IMF_plot}
\title{Plots a chosen IMF of an EEMD time series}
\usage{
IMF_plot(ts, imf)
}
\arguments{
\item{ts}{time series to manipulate}

\item{i}{the index of the IMF you want to plot}
}
\value{

}
\description{
Plots a chosen IMF of an EEMD time series
}
\examples{
library(fpp)
data(a10)
#' VVD::IMF_plot(a10,2)

}
